+++ src/main/java/org/example/bookstoreapp/common/security/JwtUtil.java
@@ -20,13 +20,11 @@ import java.util.Date;
 public class JwtUtil {
 
     private static final String BEARER_PREFIX = "Bearer ";
-    private static final long TOKEN_TIME = 60 * 60 * 1000L; // 60분
+    private static final long TOKEN_TIME = 30* 60 * 1000L; // 60분
 
     @Value("${jwt.secret.key}")
     private String secretKey;
     private SecretKey key;
-//    private final SignatureAlgorithm signatureAlgorithm = SignatureAlgorithm.HS256;
-//    변경: Base64.getDecoder() 대신 io.jsonwebtoken.io.Decoders 사용 (12.x 표준)
 
     @PostConstruct
     public void init() {
